// src/app/models/vehicle.ts
export interface Vehicle {
  id: number;
  dateAdded: string;
  makeId: number;
  mileage: number;
  colorId: number;
  conditionId: number;
  modalId: number;
  year: number;
  fuelId: number;
  transmissionId: number;
  price: string;
  description: string;
  agentName: string;
  agentPhoneNumber: string;
  agentLocation: string;
  images: any[];
  userId: number;
}